requirements.txt
groq>=0.9.0
psycopg2-binary>=2.9.0
python-jobspy>=0.1.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
streamlit>=1.30.0
pandas>=2.0.0
python-dotenv>=1.0.0
feedparser>=6.0.0
lxml>=4.9.0
openpyxl>=3.1.0
python-docx>=1.1.0
